update comment popup features, reduced size of comment popup
