+++ src/index.js
@@ -1,8 +1,7 @@
 import './style.css';
 import getMovies from './api-controller.js';
+import commentPopup from './comment.js';
 
-const commentSection = document.querySelector('.commentpop');
-const bodyfix = document.querySelector('body');
 const rederMovies = async (genresType) => {
   let movies = await getMovies();
   movies = movies.filter((item) => item.genres.indexOf(genresType) !== -1);
@@ -26,36 +25,8 @@ const rederMovies = async (genresType) => {
   }
 
   const commentButton = document.querySelectorAll('.comment-btn-div button');
-  const commentPopup = () => {
-    commentButton.forEach((button, index) => {
-      button.addEventListener('click', () => {
-        commentSection.classList.add('show');
-        bodyfix.classList.add('static');
 
-        commentSection.innerHTML = `<div class='comment-js'>
-        <div class='name-closeicon'>
-        <h2>${movies[index].name}</h2>
-        <i class="fa fa-times"></i>
-        </div>
-        <div class='comment-img'><img src='${movies[index].image.original}'></div>
-        <div class='summary'>${movies[index].summary}</div>
-        <ul>
-        <li>Genre: ${movies[index].genres.join(',')}</li>
-        <li>tvmaze: <a href="${movies[index].url}" target="_blank">details about show</a></li>
-        <li>rating: ${movies[index].rating.average}</li>
-        </ul>
-        </div>`;
-
-        const closeComment = document.querySelector('.fa-times');
-        const closebtn = () => {
-          commentSection.classList.remove('show');
-          bodyfix.classList.remove('static');
-        };
-        closeComment.addEventListener('click', closebtn);
-      });
-    });
-  };
-  commentPopup();
+  commentPopup(commentButton, movies);
 };
 
 rederMovies('Action');
