+++ src/style.css
@@ -114,7 +114,6 @@ footer p {
 
 .commentpop {
   display: none;
-  background-color: beige;
   color: black;
   position: fixed;
   top: 0;
@@ -123,12 +122,17 @@ footer p {
   overflow-y: hidden;
   overflow-x: hidden;
   justify-content: center;
+  align-items: center;
+  background-color: rgb(0, 0, 0);
+  background-color: rgba(0, 0, 0, 0.3);
 }
 
 .comment-js {
-  width: 100%;
-  height: 100%;
+  background-color: beige;
+  width: 70%;
+  height: 90%;
   overflow-y: auto;
+  border-radius: 10px;
 }
 
 .commentpop.show {
